js: Extract password_quality module; remove zxcvbn from globals.
